fix(inventory): address PR #1057 review findings
- CycleCountController interfering-movements: correct 200 schema to array
  and document the 404 ApiError body; openapi.yaml regenerated
- LocationInventoryInquiryResponse: widen quantities to long (int64) and
  drop Math.toIntExact from the inquiry service so large location totals
  cannot 500; the storage-location fact keeps its int contract with the
  conversion inside the publisher's existing per-item skip-and-warn guard
- StockSummaryDriftVerifier: push the ledger-vs-summary diff into a
  set-based native query returning only mismatching keys, so the
  scheduled job holds no table-sized state regardless of catalog and
  location cardinality; per-key WARN logging capped at 50 rows per pass

+++ pos-inventory/src/main/java/com/positivity/inventory/internal/dto/LocationInventoryInquiryResponse.java
@@ -24,12 +24,12 @@ public class LocationInventoryInquiryResponse {
             description = "Current on-hand quantity across all stock items at the location",
             example = "12",
             requiredMode = Schema.RequiredMode.REQUIRED)
-    private int onHandQuantity;
+    private long onHandQuantity;
 
     @Schema(
             description =
                     "Quantity available to promise after pending allocations. Note: reservation events are not yet factored in.",
             example = "8",
             requiredMode = Schema.RequiredMode.REQUIRED)
-    private int availableToPromiseQuantity;
+    private long availableToPromiseQuantity;
 }
